source libraries from CDNs instead of hosting in-repo
Most of them come from unpkg. Ionicons was not being used except for a
single icon in the breadcrumbs so it has been swapped to fontawesome.
Modernizr is still in the "vendor" directory in "js" because it's a
custom-ish build and modernizr doesn't publish builds anywhere a CDN
would pick them up it seems. html5shiv has been dropped as well because
it is 2020 and no one should be even close to using IE8.

closes #35